requirements.txt
alembic==1.7.5
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
PyYAML==6.0
SQLAlchemy==1.4.30
Werkzeug==2.0.2
WTForms~=2.2.1
zipp==3.7.0

langdetect~=1.0.9
PyJWT~=2.3.0
urllib3~=1.26.9
elasticsearch==8.2.0
requests~=2.27.1